+++ src/components/editor/plugins/page/PageFlowPlugin.tsx
@@ -1,0 +1,10 @@
+import { usePageFlow, type PageFlowSettings } from './usePageFlow';
+
+export function PageFlowPlugin({
+  pageHeightMm,
+  marginTopMm,
+  marginBottomMm
+}: PageFlowSettings): null {
+  usePageFlow({ pageHeightMm, marginTopMm, marginBottomMm });
+  return null;
+}
