+++ src/components/editor/plugins/page/PageObserverPlugin.tsx
@@ -1,0 +1,6 @@
+import { usePageObserver } from './usePageObserver';
+
+export function PageObserverPlugin(): null {
+  usePageObserver();
+  return null;
+}
