+++ src/components/editor/plugins/page/usePageFlow.ts
@@ -1,10 +1,10 @@
 import { useEffect, useRef } from 'react';
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
 import { $getRoot, $getSelection, $isRangeSelection } from 'lexical';
-import { $createPageNode, $isPageNode, PageNode } from '../nodes/PageNode';
+import { $createPageNode, $isPageNode, PageNode } from './PageNode';
 
 function mmToPx(mm: number): number {
-  return (mm / 25.4) * 96; // 96dpi heuristic
+  return (mm / 25.4) * 96;
 }
 
 export interface PageFlowSettings {
@@ -26,10 +26,8 @@ export function usePageFlow(settings: PageFlowSettings): void {
           return;
         }
 
-        // Work with the last page only for stability
         let last = root.getLastChild();
         if (!$isPageNode(last)) {
-          // wrap non-page children into a page (fallback)
           const page = $createPageNode();
           const children = root.getChildren();
           children.forEach((n) => {
@@ -40,8 +38,6 @@ export function usePageFlow(settings: PageFlowSettings): void {
         }
 
         const lastPage = last as PageNode;
-
-        // Measure the last page used height
         const contentHeightPx = mmToPx(
           settings.pageHeightMm - settings.marginTopMm - settings.marginBottomMm
         );
@@ -50,20 +46,15 @@ export function usePageFlow(settings: PageFlowSettings): void {
         const blocks = lastPage.getChildren();
         for (let i = 0; i < blocks.length; i++) {
           const el = editor.getElementByKey(blocks[i].getKey());
-          if (!el) {
-            continue; // DOM not ready yet
-          }
+          if (!el) continue;
           usedPx += el.offsetHeight;
         }
 
-        // If overflows (or very close), create a new page and move caret there
-        if (usedPx >= contentHeightPx - 4 /* small tolerance */) {
+        if (usedPx >= contentHeightPx - 4) {
           const next = lastPage.getNextSibling();
           if (!$isPageNode(next)) {
             const newPage = $createPageNode();
             lastPage.insertAfter(newPage);
-
-            // Move caret to the start of the new page's first block
             const firstChild = newPage.getFirstChild();
             const selection = $getSelection();
             if ($isRangeSelection(selection) && firstChild) {
@@ -77,12 +68,10 @@ export function usePageFlow(settings: PageFlowSettings): void {
     const schedule = (): void => {
       if (rafRef.current !== null) cancelAnimationFrame(rafRef.current);
       rafRef.current = requestAnimationFrame(() => {
-        // let DOM settle a tick
         setTimeout(ensureNextPageWhenFull, 0);
       });
     };
 
-    // Run once after mount
     schedule();
 
     const unregister = editor.registerUpdateListener(() => {
